git actions workflow

+++ scripts/deploy.sh
@@ -18,11 +18,12 @@ set -e
 ARTIFACT_BUCKET=$3
 
 # Find all Lambda artifacts and upload them to the S3 artifact bucket
-for i in $(ls -d __artifacts__/lambda/*.zip)
-do
-    MOD_NAME=$(basename ${i} | cut -f 1 -d '.')
-    FN_NAME="${MOD_NAME}-${2}"
-    
+# Check if the lambda zip files exist
+if ls __artifacts__/lambda/*.zip 1> /dev/null 2>&1; then
+  for file in __artifacts__/lambda/*.zip; do
+    MOD_NAME=$(basename "$file" .zip)
+    FN_NAME="${MOD_NAME}-${2}"  
+
     # Upload zip file to S3
     aws s3 cp \
       "__artifacts__/lambda/${MOD_NAME}.zip" \
@@ -35,17 +36,27 @@ do
       --s3-bucket "${ARTIFACT_BUCKET}" \
       --s3-key "lambda/${MOD_NAME}.zip" \
       --publish
-done
-
-# Upload the Reset CodeBuild Zip to the S3 artifact bucket. CodeBuild should pick this new file up on its next build.
-aws s3 cp \
-  __artifacts__/codebuild/reset.zip \
-  "s3://${ARTIFACT_BUCKET}/codebuild/reset.zip" \
-  --sse
-
-    # Delete the '__artifacts__/' directory after uploading to the s3 artifact bucket 
-    rm -rf __artifacts__
-else 
-    echo "[Error] ${FILE} does not exist yet. Run scripts/build.sh to generate it."
-    exit 1
+  done
+else
+  echo "[Error] No lambda zip files found in __artifacts__/lambda/"
+  exit 1
 fi
+
+# Check if the CodeBuild reset zip file exists
+if [ -f "__artifacts__/codebuild/reset.zip" ]; then
+  # Upload the Reset CodeBuild Zip to the S3 artifact bucket. CodeBuild should pick this new file up on its next build.
+  aws s3 cp \
+    __artifacts__/codebuild/reset.zip \
+    "s3://${ARTIFACT_BUCKET}/codebuild/reset.zip" \
+    --sse
+
+  # Delete the '__artifacts__/' directory after uploading to the s3 artifact bucket 
+  rm -rf __artifacts__
+else
+  echo "[Error] __artifacts__/codebuild/reset.zip does not exist."
+  exit 1
+fi
+
+
+
+
